requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-ollama==0.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pgvector==0.2.3
PyMuPDF==1.24.2
